+++ src/controllers/indexController.js
@@ -1,3 +1,12 @@
+// Importing config
+const config = require('../config')
+
+// Enabling colors
+config.colors.enable()
+
+/**
+ * This method will retrieve the index/root endpoint of the API
+ */
 exports.getIndex = (req, res) => {
     // TODO GET Root endpoint
     try {
@@ -10,6 +19,6 @@ exports.getIndex = (req, res) => {
         })
     }
     catch(err) {
-        throw err
+        console.log(`[nodemon] error: ${err.message}`.red)
     }
 }